+++ src/main/java/jp/crafterkina/SurviveInSociety/event/EnumEventHandler.java
@@ -8,8 +8,8 @@ package jp.crafterkina.SurviveInSociety.event;
 import jp.crafterkina.SurviveInSociety.client.eventhandler.ClientEventHandler;
 import jp.crafterkina.SurviveInSociety.event.handler.CommonEventHandler;
 import jp.crafterkina.SurviveInSociety.event.handler.FMLEventHandler;
-import jp.crafterkina.SurviveInSociety.event.handler.OreGenEventHandler;
 import jp.crafterkina.SurviveInSociety.event.handler.TerrainGenEventHandler;
+import jp.crafterkina.SurviveInSociety.world.ore.OreGenEventHandler;
 import net.minecraftforge.fml.common.FMLCommonHandler;
 import net.minecraftforge.fml.common.eventhandler.EventBus;
